+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.4.8 (in development)
 
-- Fixed loading default values for `User` creation via console
+- Added error message in `EmailController::actionTest` when email is not set (#36)
+- Fixed loading default values for `User` creation via console (#35)
 
 ## 2.4.7 (Mar 24, 2025)